separate responsibility of printing the board
