+++ src/test/java/com/ox/victorycondition/ColumnConditionTest.java
@@ -2,6 +2,7 @@ package com.ox.victorycondition;
 
 import com.ox.coordinates.BoardFieldCoordinate;
 import com.ox.core.Board;
+import com.ox.core.BoardPrinter;
 import com.ox.core.GameConfig;
 import com.ox.core.Player;
 import com.ox.language.Language;
@@ -33,6 +34,7 @@ public class ColumnConditionTest {
     Language lan = new Language();
     LanguageLoader loader = new LanguageLoader(lan, language);
     private static Board board;
+    private BoardPrinter boardPrinter;
 
     @BeforeMethod
     public void setUp() {
@@ -52,6 +54,7 @@ public class ColumnConditionTest {
         gameConfig.setLengthOfCombinationToWin(System.out::println, new Scanner(System.in)::nextLine, gameConfigValidator);
 
         board = new Board(gameConfig);
+        boardPrinter = new BoardPrinter(board);
     }
 
     private static void fillBoard(Player player, int... coordinates) {
@@ -78,7 +81,7 @@ public class ColumnConditionTest {
     public void isThereAVictory_LastSignPutOnTheEdge(int first, int second, int third, int fourth) {
         //given and when
         fillBoard(Player.X, first, second, third, fourth);
-        board.printBoardState(System.out::println);
+        boardPrinter.printBoardState(System.out::println);
         //then
         assertEquals(columnCondition.isThereAVictory(new BoardFieldCoordinate(first), board, X, gameConfig), Optional.of(X));
     }
@@ -87,7 +90,7 @@ public class ColumnConditionTest {
     public void isThereAVictory_LastSignPutInTheMiddle(int first, int second, int third, int fourth) {
         //given and when
         fillBoard(Player.X, first, second, third, fourth);
-        board.printBoardState(System.out::println);
+        boardPrinter.printBoardState(System.out::println);
         //then
         assertEquals(columnCondition.isThereAVictory(new BoardFieldCoordinate(second), board, X, gameConfig), Optional.of(X));
     }
@@ -97,7 +100,7 @@ public class ColumnConditionTest {
     public void isThereAVictory_NoVictoryInAColumn(int first, int second, int third, int fourth) {
         //given and when
         fillBoard(Player.X, first, second, third, fourth);
-        board.printBoardState(System.out::println);
+        boardPrinter.printBoardState(System.out::println);
         //then
         assertEquals(columnCondition.isThereAVictory(new BoardFieldCoordinate(first), board, O, gameConfig), Optional.empty());
     }
